Sorcerer of Dark Magic (Anime)
This card cannot be Normal Summoned or Set. This card cannot be Special Summoned except by Tributing 2 Level 6 or higher Spellcaster-Type monsters. You can negate the activation of any Trap Cards or their effect(s) and destroy them. When this card battles an opponent's monster, that monster loses 500 ATK for each Spellcaster-Type monster in your Graveyard that was sent there from the field.